_RSS/requirements.txt
boto3 >= 1.37
google-api-python-client
google-auth-httplib2
google-auth-oauthlib